# Conflicts:
#	buildconfig/appcompile/compile-customerapp.gradle
#	buildconfig/appcompile/compile-sellerapp.gradle
#	buildconfig/dependencies/dependency-features.gradle
#	customerapp/proguard-rules.txt
#	customerapp_pro/proguard-rules.txt
#	dynamic-features-customerapp_pro/base/df_base/build.gradle
#	dynamic-features-sellerapp/base/df_base_sellerapp/build.gradle
#	dynamic-features/base/df_base/build.gradle
#	features/merchant/shop_page/src/main/java/com/tokopedia/shop/pageheader/presentation/fragment/NewShopPageFragment.kt
#	libraries/applink/src/main/java/com/tokopedia/applink/FragmentConst.kt
#	libraries/applink/src/main/java/com/tokopedia/applink/FragmentDFMapper.kt
#	sellerapp/proguard-rules.txt

+++ features/content/play_broadcaster/src/main/java/com/tokopedia/play/broadcaster/view/widget/PlayRectCropImageOverlay.kt
@@ -91,13 +91,13 @@ class PlayRectCropImageOverlay @JvmOverloads constructor(context: Context, attrs
         // draw bottom black overlay
         if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
             canvas.drawRect(leftPosition,
-                    rectHeight.toFloat() - bottomDetailHeight,
+                    rectHeight - bottomDetailHeight,
                     rightPosition,
                     bottomPosition,
                     mBlackTransparentPaint)
         } else {
             canvas.drawRoundRect(leftPosition,
-                    rectHeight.toFloat() - bottomDetailHeight,
+                    rectHeight - bottomDetailHeight,
                     rightPosition,
                     bottomPosition,
                     CENTER_RECT_RADIUS,
