+++ features/logistic/editshipping/src/main/java/com/tokopedia/editshipping/ui/customproductlogistic/CPLItemAdapter.kt
@@ -69,7 +69,7 @@ class CPLItemAdapter(private val listener: CPLItemAdapterListener) :
         cplItem.forEach { courier ->
             courier.shipperProduct.forEach { product ->
                 if (product.isActive) {
-                    activatedListIds.add(product.shipperProductId)
+                    activatedListIds.add(product.shipperProductId.toInt())
                 }
             }
         }
@@ -81,7 +81,7 @@ class CPLItemAdapter(private val listener: CPLItemAdapterListener) :
         cplItem.forEach { courier ->
             courier.shipperProduct.forEach { product ->
                 if (product.isActive) {
-                    activatedListIds.add(product.shipperProductId)
+                    activatedListIds.add(product.shipperProductId.toInt())
                 }
             }
         }
@@ -93,7 +93,7 @@ class CPLItemAdapter(private val listener: CPLItemAdapterListener) :
         cplItem.forEach { courier ->
             courier.shipperProduct.forEach { product ->
                 if (!product.uiHidden) {
-                    listShipperShown.add(product.shipperProductId)
+                    listShipperShown.add(product.shipperProductId.toInt())
                 }
             }
         }
