+++ features/logistic/localizationchooseaddress/src/main/java/com/tokopedia/localizationchooseaddress/domain/response/GetDefaultChosenAddressGqlResponse.kt
@@ -24,7 +24,7 @@ data class GetDefaultChosenAddressResponse(
 
 data class DefaultChosenAddressData(
         @SerializedName("addr_id")
-        var addressId: Int = 0,
+        var addressId: Long = 0,
         @SerializedName("receiver_name")
         var receiverName: String = "",
         @SerializedName("addr_name")
