+++ features/logistic/logisticCommon/src/main/java/com/tokopedia/logisticCommon/domain/usecase/GetAddressList.kt
@@ -23,20 +23,20 @@ const val PARAM_ADDRESS_USECASE: String = "input"
 class GetAddressCornerUseCase
 @Inject constructor(@ApplicationContext val context: Context, val usecase: GraphqlUseCaseInterface, val mapper: AddressCornerMapper) {
 
-    fun execute(query: String, prevState: Int?, localChosenAddrId: Int?, isWhitelistChosenAddress: Boolean): Observable<AddressListModel> =
+    fun execute(query: String, prevState: Int?, localChosenAddrId: Long?, isWhitelistChosenAddress: Boolean): Observable<AddressListModel> =
             this.getObservable(query = query, page = 1, isAddress = true, isCorner = false, limit = 10,
                     prevState = prevState, localChosenAddrId = localChosenAddrId, isWhitelistChosenAddress)
 
-    fun getAll(query: String, prevState: Int, localChosenAddrId: Int, isWhitelistChosenAddress: Boolean): Observable<AddressListModel> =
+    fun getAll(query: String, prevState: Int, localChosenAddrId: Long, isWhitelistChosenAddress: Boolean): Observable<AddressListModel> =
             this.getObservable(query = query, page = 1, isAddress = true, isCorner = false, limit = 0,
                     prevState = prevState, localChosenAddrId = localChosenAddrId, isWhitelistChosenAddress = isWhitelistChosenAddress)
 
-    fun loadMore(query: String, page: Int, prevState: Int?, localChosenAddrId: Int?, isWhitelistChosenAddress: Boolean): Observable<AddressListModel> =
+    fun loadMore(query: String, page: Int, prevState: Int?, localChosenAddrId: Long?, isWhitelistChosenAddress: Boolean): Observable<AddressListModel> =
             this.getObservable(query = query, page = page, isAddress = true, isCorner = false, limit = 10,
                     prevState = prevState, localChosenAddrId = localChosenAddrId, isWhitelistChosenAddress = isWhitelistChosenAddress)
 
     private fun getObservable(query: String, page: Int, isAddress: Boolean, isCorner: Boolean, limit: Int,
-                              prevState: Int?, localChosenAddrId: Int?, isWhitelistChosenAddress: Boolean):
+                              prevState: Int?, localChosenAddrId: Long?, isWhitelistChosenAddress: Boolean):
             Observable<AddressListModel> {
         val request = AddressRequest(searchKey = query, page = page, showAddress = isAddress,
                 showCorner = isCorner, limit = limit, whitelistChosenAddress = isWhitelistChosenAddress, previousState = prevState,
